Revert "automatically expand watches after adding a watch"
The parent commit fixed automatic expansion of a section "Auto" of the
Variables tool view in VariableTree(). An instance of VariableTree is
created only when a Variables tool view widget (VariableWidget) is
created. Unless the user removes the Variables tool view from all
sublime areas then re-adds it, the section "Auto" is expanded only once
after each KDevelop start, while there are zero watches. The UI does not
indicate the expanded state of a tree item that has zero children in any
way.

A similar feature - an automatic expansion of the section "Auto" when a
first watch is added - is implemented in Watches::add() and
Watches::addFinishResult(). This mostly duplicating feature was
implemented to address Bug 278652. But I do not understand why that bug
had been reported in the first place, because the aforementioned feature
in VariableTree() should have been working correctly at that time (see
the parent commit message). Perhaps some other bug had prevented the
feature from working as expected at that time.

The automatic expansion in Watches::add() and Watches::addFinishResult()
does not work and probably has never worked as intended, because
TreeItem::setExpanded() is normally invoked in response to item
expansion in the UI by the user, and so calling this function does not
actually expand the item in the UI. But these calls to setExpanded()
enable updates of the values of watched variables even if the section
"Auto" that contains them is collapsed, which is a waste of CPU time.

The non-working feature is inferior to the aforementioned feature in
VariableTree() because automatically expanding the section "Auto"
whenever it becomes nonempty would override an explicit collapsing by
the user. In case of zero watches and repeated stepping out of non-void
functions whose return values the user is not interested in, the section
"Auto" would be annoyingly expanded in Watches::addFinishResult() no
matter how many times the user collapsed it.

While the reverted feature does not work as intended, it affects
debugger unit tests. Reverting it causes the following two failures of
test_gdb as well as two analogous failures of test_lldb:
FAIL!  : KDevMI::GDB::GdbTest::testVariablesWatches() 'Comparison of data' returned FALSE. ('0' didn't match expected '1' in kdevelop/plugins/debuggercommon/tests/debuggertestbase.cpp:1756)
   Loc: [kdevelop/plugins/debuggercommon/tests/debuggertestbase.cpp(1756)]
FAIL!  : KDevMI::GDB::GdbTest::testVariablesChanged() Compared values are not the same
   Actual   (variableController->autoUpdate())  : 0
   Expected (IVariableController::UpdateWatches): 2
   Loc: [kdevelop/plugins/debuggercommon/tests/debuggertestbase.cpp(2092)]

Fix the test failures by expanding the section "Auto" in
DebuggerTestBase::init() to match the behavior of the KDevelop UI, and
adjusting the affected test functions accordingly.

This reverts commit 978a70d51f8d82867a33041d89630c5f639c58ad.