shell: don't try to show a message box in NoUi mode
During startup KDevelop displays an error message in a QMessageBox or a
KMessageBox if no plugins are found or the session cannot be locked
(e.g. because the same executable is already running). An attempt to
show a QMessageBox or a KMessageBox in a mode Core::Setup::NoUi prints a
message like the following and aborts:
QFATAL : TestDocumentchangeset::initTestCase() QWidget: Cannot create a QWidget without QApplication

Print the error message via qCCritical() in the NoUi mode instead - to
actually deliver the message to the user and to replace the irrelevant
failure reason with the following dedicated one:
QFATAL : TestDocumentchangeset::initTestCase() could not initialize the core