Disable action Stop on Change for return value variable
Triggering a context menu action Stop on Change while a return value
variable is selected in the Variables tool view sets a watchpoint on the
variable's expression, e.g. $1. The value of this expression never
changes, so the watchpoint is never hit.

Disable the action Stop on Change while a return value variable is
selected so that the user does not wonder why triggering it has no
useful effect.

Return early from VariableTree::stopOnChange() if a return value
variable is selected to preemptively optimize in case the action Stop on
Change gains a shortcut and becomes triggerable without a context menu.