VariableTree: explain return value variable in context menu title
The displayed name $ret and the context menu title, e.g. $1, of a return
value variable are not self-explanatory. Explain what a return value
variable represents in its context menu title.

Explaining a return value variable in a tooltip for its name would also
be helpful. But implementing this requires either storing more data in a
Variable (which such a trivial feature does not warrant) or accessing
the instance of Watches in Variable::data(). Accessing global data from
the function data() of an item model is crash prone.