Extract MIVariable::setValueToOptionalValueFieldOf() and optimize
A reference to MI::Value will be passed to the extracted function in the
next commit. Therefore, do not require a reference to a derived class
ResultRecord that is passed to the extracted function by its two callers
in this commit.

Optimize the extracted function and code that calls it slightly:
* deduplicate a string literal "value";
* introduce local raw pointer variables to eliminate redundant weak
  pointer accesses in QPointer.