debuggercommon: don't pass --thread and --frame to -var-set-format
An MI command -var-set-format takes the name of an existing variable
object as an argument. Consequently, neither GDB/MI nor LLDB-MI need a
thread and frame context for this command. However, GDB/MI (but not
LLDB-MI) always verifies that the thread and frame arguments are valid.
So if an ID of a nonexistent thread is accidentally passed to
-var-set-format, the GDB/MI command fails. For example:
    (gdb) 118-var-set-format --thread 2 --frame 0 var21 hexadecimal
    118^error,msg="Invalid thread id: 2"
Such a failure would desynchronize Variable::format() and the format of
the corresponding GDB variable object.

Do not pass the unused --thread and --frame options to -var-set-format
in order to eliminate the unnecessary risk of desynchronization, as well
as slightly improve performance by doing less work both in KDevelop and
in GDB/MI.