Deduplicate formatChanged() of MIVariable and LldbVariable
Introduce a new virtual helper function
MIVariable::handlerOfSetFormatCommand() in order to deduplicate
equivalent code in a member function formatChanged() of MIVariable and
its derived class LldbVariable.

The only consequential distinction eliminated by the deduplication is
that MIVariable casts to MIVariable* via dynamic_cast and LldbVariable -
via qobject_cast. This distinction was introduced almost certainly
unintentionally in 8dc24a570ac21832a3b738a46e59d318184893da. The two
casts *should* produce identical results. However, kdevdebuggercommon is
a static library as a FIXME comment in debuggercommon/CMakeLists.txt
documents. Consequently, a qobject_cast to MIVariable* of a GdbVariable*
in kdevlldb or of a LldbVariable* in kdevgdb returns a null pointer, and
a dynamic_cast - a nonnull pointer. A fix of a crash reported in
Bug 509612 would prevent reuse of variables of incompatible types and
render the distinction unimportant. For now, an assertion that the
dynamic_cast returns a non-null pointer is less likely to fail, so keep
it.

Replace QStringLiteral(" %1 %2 ").arg(V, F) with
`V + QLatin1Char{' '} + F` in order to remove redundant spaces from the
command string and optimize.