MIVariable: don't send -var-set-format with empty variable name
Sending an MI command without a mandatory argument, i.e. the name of a
variable object, always fails. The replies to an MI command
"28-var-set-format  octal" of GDB/MI and LLDB-MI respectively:
28^error,msg="-var-set-format: Usage: NAME FORMAT."
28^error,msg="Command 'var-set-format'. Command Args. Validation failed. Mandatory args not found: format-spec. Not all arguments or options were recognised:  "

So soon as a MIVariable is attached and its varobj() becomes nonempty,
CreateVarobjHandler::handle() adds a -var-set-format command to
synchronize a format selected in the UI with the debugger (unless the
current format is Natural, which matches the debugger's default).